fix: ensure delete method in reports that have been reported more than once now diminishes in one the reportQuantity instead of deleting the full record

+++ api/src/main/java/com/a3bradesco/api/controllers/CellphoneReportController.java
@@ -61,14 +61,24 @@ public class CellphoneReportController {
 
     @DeleteMapping("/{cellphone}")
     public ResponseEntity<String> deleteReport(@PathVariable String cellphone){
-        cellphoneReportService.deleteById(cellphone);
-        CellphoneReport isDeleted = cellphoneReportService.findById(cellphone);
-        if(isDeleted == null){
-            return ResponseEntity.ok("Denúncia retirada com sucesso!");
-        } else {
-            return ResponseEntity.badRequest().build();
-        }
-        
+    CellphoneReport currentDatabaseReport = cellphoneReportService.findById(cellphone);
+
+    if(currentDatabaseReport == null) {
+        return ResponseEntity.notFound().build();
     }
-    
+    if(currentDatabaseReport.getReportQuantity() <= 1){
+        cellphoneReportService.deleteById(cellphone);
+        return ResponseEntity.ok("Denúncia retirada com sucesso!");
+    } else {
+        CellphoneReport newDatabaseReport = 
+        new CellphoneReport(
+            currentDatabaseReport.getCellphone(), 
+            currentDatabaseReport.getReportQuantity() - 1, 
+            currentDatabaseReport.getLastTimeReported());
+
+        cellphoneReportService.insert(newDatabaseReport);
+
+        return ResponseEntity.ok("Denúncia retirada com sucesso!");
+        }
+    }
 }
